Make sure that the wheel is spinning with a constant speed

+++ Tasks/MessageManager.h
@@ -16,7 +16,7 @@ typedef struct {
 
 // Message structs
 typedef struct {
-	uint8_t speed;
+	uint16_t speed;
 	uint8_t direction;
 } Message_Speed_t;
